pyproject.toml
[tool.poetry]
name = "exam"
version = "0.1.0"
description = ""
authors = ["Aucke <aucke.bos97@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
tensorflow-gpu = "2.2.0"
tensorboard = "2.2.0"
pandas = "^1.2.3"
scikit-learn = "^0.24.1"
numpy = "^1.20.1"
matplotlib = "^3.3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
